wrapper-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
requests==2.31.0
python-multipart==0.0.6
pydantic-settings==2.1.0
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
pyyaml==0.23.3
respx==0.20.2
